Extract specs/ subpackage from session_service/
Mirror the existing provisioning/ and backends/ layout: SessionSpec et al
move to specs/types.py and build_spec_for_session moves to specs/factory.py,
with specs/__init__.py re-exporting both.

Public absolute import paths (agent_on_demand.session_service.specs) keep
working unchanged; the only callers that needed updating were the relative
imports inside session_service/ and the two test files that imported
spec_factory directly.
