+++ tests/test_spec_factory.py
@@ -45,7 +45,7 @@ import uuid
 from types import SimpleNamespace
 
 from agent_on_demand.runtimes import RUNTIMES
-from agent_on_demand.session_service.spec_factory import build_spec_for_session
+from agent_on_demand.session_service.specs import build_spec_for_session
 
 _UNSET = object()
